changes the structure and link in code. You should change the last problem with IconName
