--- my-react-app/components/Section/Section.jsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import css from './Section.module.css';
-
-export default function Section({ children }) {
-  return <section className={css.container}>{children}</section>;
-}
