--- my-react-app/components/Section/Section.module.css
@@ -1,7 +1,0 @@
-.container {
-    margin-left: auto;
-    margin-right: auto;
-    padding: 15px;
-    width: calc(100%-60px);
-}
-
