+++ my-react-app/src/App.jsx
@@ -1,22 +1,22 @@
 import './App.css';
-import {Navigation} from '../components/Navigation'
-import {Header_banner} from '../components/Header_banner'
-import {Categories} from '../components/Categories'
-import {Discount_card}  from '../components/Discount_card'
-import {Sale} from '../components/Sale'
-import {Contact} from '../components/Contact'
-import {Map} from '../components/Map'
+// import {Navigation} from '../components/Navigation'
+// import {Header_banner} from '../components/Header_banner'
+// import {Categories} from '../components/Categories'
+// import {Discount_card}  from '../components/Discount_card'
+// import {Sale} from '../components/Sale'
+import Contact from '../src/components/Contact/Contacts'
+import Map from '../src/components/Map/Map'
 
 export const App = () => {
   return (
     <>
-      <Navigation />
+      {/* <Navigation />
       <Header_banner />
       <Categories />
       <Discount_card />
-      <Sale />
-      <Contact />
-      <Map />
+      <Sale /> */}
+      {/* <Contact /> */}
+      {/* <Map /> */}
     </>
   );
 }
