+++ my-react-app/src/components/Section/Section.module.css
@@ -1,0 +1,7 @@
+.container {
+    margin-left: auto;
+    margin-right: auto;
+    padding: 15px;
+    width: calc(100%-60px);
+}
+
